+++ src/components/TodoComponents/Item.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Item = props => {
+	return <div className='Item'>{props.item.task}</div>;
+};
+
+export default Item;
