+++ src/components/TodoComponents/Todo.js
@@ -1,0 +1,43 @@
+//data
+export const tasks = [
+	{
+		task: 'Clean Kitchen',
+		id: 123,
+		completed: false
+	},
+	{
+		task: 'Do Laundry',
+		id: 124,
+		completed: false
+	},
+	{
+		task: 'Take Out Trash',
+		id: 125,
+		completed: false
+	},
+	{
+		task: 'Make Dinner',
+		id: 126,
+		completed: false
+	},
+	{
+		task: 'Get Groceries',
+		id: 127,
+		completed: false
+	},
+	{
+		task: 'Make Dinner',
+		id: 128,
+		completed: false
+	},
+	{
+		task: 'Put Kids To Bed',
+		id: 129,
+		completed: false
+	},
+	{
+		task: 'Chillax',
+		id: 130,
+		completed: false
+	}
+];
